Adds basic Config module
The module handles all logic related to the Config struct, which is
in charge of storing the paths to the files from which the program
can read contents to perform a given operation.

Currently, the struct `Config` has three fields: file_1 and file_2, both
of type `PathBuf` as well as a third field, called data, of type
`String`, which represents the buffer where the contents of the file
that will be read by program will be loaded.

Two associate functions (new and set_file_paths) are in charge of
instantiating a new instance of the struct and of settign the paths
to the files used by the program.