Removes field in Config + Adds associate functions and Args module
The data field in Config, which was initialised as an empty `String` and
represented the buffer for the contents of the file that was to be read
was removed due to the fact that it simply did not make sense for the
struct to hold said data.

Four new associate functions were added: `get_file_1_path` and
`get_file_2_path`, which return a `&Path` to a file; `load_data`, which
loads the data from a given file and takes a &Path object as an arg and
`run`, which matches the arg with which the program is run to a given
action: load file 1 or file 2.

This last function also made the holding of the data contents within the
Config struct redundant. It returns an instance `Option<String>` where
`Some(String)` represents the contents of the file and `None` stands
for a file with empty contents, which should never be the case for the
purposes of this program.

The returned String should be used to initialised another struct that
best semantically represents the logic behind the code.

All functions include tests and documentation.

Args module

The args module introduces the Args enum, which represents the possible
args that can be used to run the program. The args are passed to
`Config.run()`