Adds data module and tests
The data module handles the logic regarding the contents extracted
from the file used by the program. It holds only two fields: data, which
represents the contents of the file and word_count, which stands for
the amount of words extracted from the file.

Right now, the data field is represented by a String; however, this is
likely to be replaced by some form of `Mutex` since the data must be
passed to different threads for further processing, although this is
for now only a theory.