Adds Task module + Changes Config::run logic + Expands Data module
Task module

The task module handles the logic behind choosing what operation will
be performed on the chosen file. It does so by matching the argument
with which the program is run to a function. The pointer to the chosen
function is stored with the operation field of the struct.

There program displays two possible operations: target_p, which changes
any word that begins with the character `p` to another given word and
target_s, which finds all occurrences of the character `s` and replaces
them for a given String.

Replacements for the words starting with `p` and for character `s` are
defined at the top of the module.

Changes to Config::run

In order to best semantically convey what the code is trying to express,
the return type of the function was changed to Option<(Data, Task),
since by the time the config finished running, the program should know
what the task that is to be performed is as well as the data that
will be used for the operation.

Additionally, a test required changing since previously Config::run was
also used to test whether the files were empty or not: now, however,
Config::load_data is in charge of testing that functionality.

At the moment, there is no test for Config::run.

Data module

The data module was expanded with one extra associate function
(get_word_count), which simply returns the number of words that were
extracted from the file as a u32.

Other minor changes:
Updates to documentation and README file